Include deleteTask function test using Jest.
